added pi_simple_question module and Quiz module to do worksheet and quiz activities.
NOTE I need to update the authoring interface to allow an author to add these quiz activities in a different way to what we do so far.  I.e. the author needs to be able to set the node_type to view and then add info to  pi_activity_node table so that the quiz is picked up.